--- Recipes/Recipes/Auth Screen/AuthModel.swift
@@ -1,19 +1,0 @@
-//
-//  AuthModel.swift
-//  Recipes
-//
-//  Created by Vermut xxx on 28.02.2024.
-//
-
-import UIKit
-/// Модель
-class AuthModel: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-    
-    
-}
